Added functionality to create an array of user objects
The next part of this app is the data aspect, where we will need to retrieve the data in the input fields in the register tab when a user clicks the submit button of the form. First initialize an empty array called “users”, which will be used to store your list of user objects that you will be dynamically creating using the HTML form in the register tab. Remember when the submit button is clicked, you must change tabs to the usersTab as well.

We will then need to create a function with the name of “createUser” that uses this data, creates a JavaScript object with the variable name of “newUser” and finally adds it to the “users” Array.